Reads in CSV and checks fileName, and column datatypes.
Created the example csv files in exampleFiles/.
In FirstTest.java it reads in 3 arguments from the command-line: input file name, input file type, and the input file path. The input file name must match the actual file name give in the path. If the file is empty, no columns will be found.